+++ src/views/ResidentView/ResidentInformation/DocumentInformation/GenerateContractModal/GenerateContractModal.tsx
@@ -46,7 +46,9 @@ function GenerateContractModal(props: GenerateContractModalProps) {
   const selectedResident = useRecoilValue(
     residentViewSelectedResidentState,
   ) as Resident;
-  const { addDocument, extendRentInformation } = useResident(selectedResident.id);
+  const { addDocument, extendRentInformation } = useResident(
+    selectedResident.id,
+  );
 
   const {
     formInput,
@@ -69,13 +71,17 @@ function GenerateContractModal(props: GenerateContractModalProps) {
           contractStart: values.contractStart,
           landlord,
           property,
-          resident: applyHistoryToResident(residentForContract, values.contractStart),
+          resident: applyHistoryToResident(
+            residentForContract,
+            values.contractStart,
+          ),
         })
         .then((documentId) => {
           addDocument({
             id: documentId,
             type: DocumentType.Contract,
-            date: values.contractStart,
+            creationDate: new Date(),
+            subjectDate: values.contractStart,
             name: `Mietvertrag ${values.contractStart.toString()}`,
           });
           props.onClose();