+++ src/views/ResidentView/ResidentInformation/DocumentInformation/UploadDocumentModal/UploadDocumentModal.spec.tsx
@@ -49,7 +49,9 @@ describe('UploadDocumentModal', () => {
       });
     }
 
-    mockedIpcAPIFunctions.fileSystem.selectFile.mockResolvedValueOnce(inputValues.file);
+    mockedIpcAPIFunctions.fileSystem.selectFile.mockResolvedValueOnce(
+      inputValues.file,
+    );
 
     act(() => {
       input(baseElement.querySelector('#name'), inputValues.name ?? '');
@@ -60,7 +62,9 @@ describe('UploadDocumentModal', () => {
       fireEvent.click(baseElement.querySelector('#file')!);
     });
 
-    await waitFor(() => expect(mockedIpcAPIFunctions.fileSystem.selectFile).toHaveBeenCalledTimes(1));
+    await waitFor(() => expect(mockedIpcAPIFunctions.fileSystem.selectFile).toHaveBeenCalledTimes(
+      1,
+    ));
   }
 
   function submitForm() {
@@ -114,10 +118,13 @@ describe('UploadDocumentModal', () => {
     // Arrange
     const expectedDocument: Omit<LinkedDocument, 'id'> = {
       name: validInputValues.name,
-      date: validInputValues.date,
+      creationDate: validInputValues.date,
+      subjectDate: validInputValues.date,
       type: DocumentType.Contract,
     };
-    mockedIpcAPIFunctions.persistence.uploadDocument.mockResolvedValue(undefined);
+    mockedIpcAPIFunctions.persistence.uploadDocument.mockResolvedValue(
+      undefined,
+    );
 
     await inputToForm(validInputValues);
 
@@ -127,8 +134,12 @@ describe('UploadDocumentModal', () => {
     await waitFor(() => expect(useResidentMock.addDocument).toHaveBeenCalledTimes(1));
 
     // Assert
-    expect(mockedIpcAPIFunctions.persistence.uploadDocument).toHaveBeenCalledTimes(1);
-    expect(mockedIpcAPIFunctions.persistence.uploadDocument).toHaveBeenLastCalledWith(
+    expect(
+      mockedIpcAPIFunctions.persistence.uploadDocument,
+    ).toHaveBeenCalledTimes(1);
+    expect(
+      mockedIpcAPIFunctions.persistence.uploadDocument,
+    ).toHaveBeenLastCalledWith(
       validInputValues.file,
       expect.stringMatching(/^.*\.txt$/),
       {