+++ src/views/ResidentView/ResidentView.spec.tsx
@@ -80,13 +80,13 @@ describe('ResidentView', () => {
     .addDocument(
       new LinkedDocumentBuilder()
         .withName('Document 1')
-        .withDate(new Date(2024, 5, 9))
+        .withSubjectDate(new Date(2024, 5, 9))
         .build(),
     )
     .addDocument(
       new LinkedDocumentBuilder()
         .withName('Document 2')
-        .withDate(new Date(2024, 5, 8))
+        .withSubjectDate(new Date(2024, 5, 8))
         .build(),
     )
     .build());